Fix UI positioning and index player movement
added horizontal layout to the Tool UI canvas to organize the buttons and set canvas scaler to scale with screen size so button stays the same size and position under different screen ratio, same for PauseCanvas. change canvas setting of PauseCanvas to screen space - camera with main cam to make sure the pause menu is covering the main cam screen.